refactor(InputView): move const val to view
